requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
joblib==1.3.2
nltk==3.8.1
requests==2.31.0
googletrans==4.0.0rc1
langdetect==1.0.9
gunicorn==21.2.0
Werkzeug==3.0.1